Controller
Haciendo el controlador y rutas del modelo de servicio

+++ app/Http/Controllers/HotelController.php
@@ -12,7 +12,7 @@ use Api\Direccion;
 use Api\Galeria;
 use Api\Requisitos;
 use Api\DocumentosSolicitar;
-
+use Log;
 use Session;
 use Redirect;
 class HotelController extends Controller
@@ -27,18 +27,9 @@ class HotelController extends Controller
     {
         
         $hotel=Hotel::all();        
-        return $hotel;
+        return response()->json($hotel,200);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
 
     /**
      * Store a newly created resource in storage.
@@ -49,88 +40,77 @@ class HotelController extends Controller
     public function store(Request $request)
     {
         //
-        $direccion= new Direccion;
-        $direccion->pais= $request->pais;
-        $direccion->ciudad= $request->idciudad;
-        $direccion->codigoPostal= $request->codpostal;
-        $direccion->idPais= $request->idpais;
-        $direccion->calle= $request->idcalle;
-        $direccion->longitud= $request->longitud;
-        $direccion->latitud= $request->latitud;
-        $direccion->save();
-
-        $hotel = new Hotel;
-        $hotel->direccion_id=$direccion->id;
-        $hotel->nombre=$request->nombhotel;
-        $hotel->smallName=$request->smallname;
-        $hotel->descripcion=$request->descripcion;      
-        $hotel->rating=$request->rating;
-        $hotel->activo=true;        
-        $hotel->save();     
-
-        
-
-        
-        if($request->hasFile('ruta'))
-        {
-            
-            foreach($request->file('ruta') as $media)
-            {
-                    $galeria=new Galeria;
-                    $galeria->ruta = $media->getClientOriginalName();
-                    $galeria->hotel()->associate($hotel);
-                    \Storage::disk('local_galeria')->put($galeria->ruta,  \File::get($media));              
-                    $galeria->save();
-            
+        try{
                 
-            }
+                $direccion= Direccion::create($request->all());                    
+                $hotel = Hotel::create($request->all());                                
+
+                
+                if($request->hasFile('ruta'))
+                {
+                    
+                    foreach($request->file('ruta') as $media)
+                    {
+                            $galeria=new Galeria;
+                            $galeria->ruta = $media->getClientOriginalName();
+                            $galeria->hotel()->associate($hotel);
+                            \Storage::disk('local_galeria')->put($galeria->ruta,  \File::get($media));              
+                            $galeria->save();
+                    
+                        
+                    }
 
 
-            
-        }
-        
-
-        $disponible= $this->multiexplode(array(","),$request->servicios_disponibles);
-        $destacado= $this->multiexplode(array(","),$request->servicios_destacados);  
-        
-        
-        $igual=true;         
-        for ($i=0; $i < count($disponible)-1 ; $i++) {         
-            for ($j=0; $j < count($destacado)-1 ; $j++) { 
-                if($disponible[$i]==$destacado[$j] )
-                {    
-                    $igual=false;
-                    break;               
+                    
                 }
-            }
-            if(!$igual)
-            {
-              $hotel->servicios()->attach($disponible[$i],['destacado' => true, 'disponible'=>true]);
-            }
-            else{
-              $hotel->servicios()->attach($disponible[$i],['destacado' => false, 'disponible'=>true]);
-            }
-            
-        }
-        $diff=true;
-        for ($i=0; $i < count($destacado)-1 ; $i++) {         
-            for ($j=0; $j < count($disponible)-1 ; $j++) { 
-                if($destacado[$i]==$disponible[$j] )
-                {   
-                    $diff=false;         
-                    break;
-                }               
-            }
-            if($diff)
-            {
-                $hotel->servicios()->attach($destacado[$i],['destacado' => true, 'disponible'=>false]);
-            }            
-        }
+                
+
+                $disponible= $this->multiexplode(array(","),$request->servicios_disponibles);
+                $destacado= $this->multiexplode(array(","),$request->servicios_destacados);  
+                
+                
+                $igual=true;         
+                for ($i=0; $i < count($disponible)-1 ; $i++) {         
+                    for ($j=0; $j < count($destacado)-1 ; $j++) { 
+                        if($disponible[$i]==$destacado[$j] )
+                        {    
+                            $igual=false;
+                            break;               
+                        }
+                    }
+                    if(!$igual)
+                    {
+                      $hotel->servicios()->attach($disponible[$i],['destacado' => true, 'disponible'=>true]);
+                    }
+                    else{
+                      $hotel->servicios()->attach($disponible[$i],['destacado' => false, 'disponible'=>true]);
+                    }
+                    
+                }
+                $diff=true;
+                for ($i=0; $i < count($destacado)-1 ; $i++) {         
+                    for ($j=0; $j < count($disponible)-1 ; $j++) { 
+                        if($destacado[$i]==$disponible[$j] )
+                        {   
+                            $diff=false;         
+                            break;
+                        }               
+                    }
+                    if($diff)
+                    {
+                        $hotel->servicios()->attach($destacado[$i],['destacado' => true, 'disponible'=>false]);
+                    }            
+                }
         
 
 
-        Session::flash('messageok','El Hotel ha sido creado correctamente');    
-        return Redirect::to('hoteles');
+            return response()->json(['status'=>true, 'message'=>'Muchas Gracias'], 200);
+        }
+        catch(\Exception $e)
+        {
+            Log::critical("No se puede agregar el Hotel:{$e->getCode()}, {$e->getLine()}, {$e->getMessage()} ");
+            return response("Alguna cosa esta mal", 500);
+        }
     }
 
     public function multiexplode ($delimiters,$string) {
@@ -147,26 +127,21 @@ class HotelController extends Controller
      */
     public function show($id)
     {
-        //
+        try{
+            $hotel = Hotel::find($id);
+            if (!$hotel) {
+                return response("No existe el hotel", 404);
+            }            
+            return response()->json($hotel, 200);
+        }
+        catch(\Exception $e)
+        {
+            Log::critical("No se puede mostrar el Hotel :{$e->getCode()}, {$e->getLine()}, {$e->getMessage()} ");
+            return response("Alguna cosa esta mal", 500);
+        } 
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
-          $hotel = Hotel::find($id);
-        //$collection = Collection::make($hotel);   
-
-        return response()->json(
-            $hotel->toArray()
-            ); 
-    }
-
+ 
     /**
      * Update the specified resource in storage.
      *
@@ -176,86 +151,76 @@ class HotelController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $id=$request->idhotel;      
-        $hotel = Hotel::find($id);
-        $direccion= Direccion::find($hotel->direccion_id);
-        $direccion->pais= $request->pais;
-        $direccion->ciudad= $request->idciudad;
-        $direccion->codigoPostal= $request->codpostal;
-        $direccion->idPais= $request->idpais;
-        $direccion->calle= $request->idcalle;
-        $direccion->longitud= $request->latitud;
-        $direccion->latitud= $request->longitud;
-        $direccion->save();
-        
-        //$hotel->direccion_id=$direccion->id;
-        $hotel->nombre=$request->nombhotel;
-        $hotel->smallName=$request->smallname;
-        $hotel->descripcion=$request->descripcion;      
-        $hotel->rating=$request->rating;
-        $hotel->activo=true;    
-        $hotel->save();     
+        try{
+            $id=$request->idhotel;  
+            $hotel = Hotel::find($id);
+            $direccion= Direccion::find($hotel->direccion_id);
 
+            $direccion->fill($request->all()); 
+            $hotel->fill($request->all()); 
         
-
-        
-        if($request->hasFile('ruta'))
-        {
-            
-            foreach($request->file('ruta') as $media)
-            {
-                    $galeria=new Galeria;
-                    $galeria->ruta = $media->getClientOriginalName();
-                    $galeria->hotel()->associate($hotel);
-                    \Storage::disk('local_galeria')->put($galeria->ruta,  \File::get($media));              
-                    $galeria->save();
-            
-                
-            }           
-        }
-        
-
-        $disponible= $this->multiexplode(array(","),$request->servicios_disponibles);
-        $destacado= $this->multiexplode(array(","),$request->servicios_destacados);  
-        
-        
-        $igual=true;         
-        for ($i=0; $i < count($disponible)-1 ; $i++) {         
-            for ($j=0; $j < count($destacado)-1 ; $j++) { 
-                if($disponible[$i]==$destacado[$j] )
-                {    
-                    $igual=false;
-                    break;               
+                if($request->hasFile('ruta'))
+                {
+                    
+                    foreach($request->file('ruta') as $media)
+                    {
+                            $galeria=new Galeria;
+                            $galeria->ruta = $media->getClientOriginalName();
+                            $galeria->hotel()->associate($hotel);
+                            \Storage::disk('local_galeria')->put($galeria->ruta,  \File::get($media));              
+                            $galeria->save();
+                    
+                        
+                    }           
                 }
-            }
-            if(!$igual)
-            {
-              $hotel->servicios()->sync([$disponible[$i],['destacado' => true, 'disponible'=>true]]);
-            }
-            else{
-              $hotel->servicios()->sync([$disponible[$i],['destacado' => false, 'disponible'=>true]]);
-            }
-            
-        }
-        $diff=true;
-        for ($i=0; $i < count($destacado)-1 ; $i++) {         
-            for ($j=0; $j < count($disponible)-1 ; $j++) { 
-                if($destacado[$i]==$disponible[$j] )
-                {   
-                    $diff=false;         
-                    break;
-                }               
-            }
-            if($diff)
-            {
-                $hotel->servicios()->sync([$destacado[$i],['destacado' => true, 'disponible'=>false]]);
-            }            
-        }
         
 
-        Session::flash('messageok','El Hotel ha sido actualizado correctamente');   
+            $disponible= $this->multiexplode(array(","),$request->servicios_disponibles);
+            $destacado= $this->multiexplode(array(","),$request->servicios_destacados);  
+        
+        
+            $igual=true;         
+            for ($i=0; $i < count($disponible)-1 ; $i++) {         
+                for ($j=0; $j < count($destacado)-1 ; $j++) { 
+                    if($disponible[$i]==$destacado[$j] )
+                    {    
+                        $igual=false;
+                        break;               
+                    }
+                }
+                if(!$igual)
+                {
+                  $hotel->servicios()->sync([$disponible[$i],['destacado' => true, 'disponible'=>true]]);
+                }
+                else{
+                  $hotel->servicios()->sync([$disponible[$i],['destacado' => false, 'disponible'=>true]]);
+                }
+                
+            }
+            $diff=true;
+            for ($i=0; $i < count($destacado)-1 ; $i++) {         
+                for ($j=0; $j < count($disponible)-1 ; $j++) { 
+                    if($destacado[$i]==$disponible[$j] )
+                    {   
+                        $diff=false;         
+                        break;
+                    }               
+                }
+                if($diff)
+                {
+                    $hotel->servicios()->sync([$destacado[$i],['destacado' => true, 'disponible'=>false]]);
+                }            
+            }
+        
 
-        return Redirect::to('hoteles');
+            return response()->json(['status'=>true, 'message'=>'Muchas Gracias'], 200);
+
+        }
+        catch(\Exception $e)
+        {
+            Log::critical("No se puede agregar el Hotel:{$e->getCode()}, {$e->getLine()}, {$e->getMessage()} ");
+            return response("Alguna cosa esta mal", 500);
+        }
     }
 
     /**
@@ -266,14 +231,24 @@ class HotelController extends Controller
      */
     public function destroy($id)
     {
-        $hotel = Hotel::find($id);
-        $hotel->delete();
-        
-        foreach ($hotel->galeria as $galeria) {        
+        try{
+            $hotel = Hotel::find($id);
+            if (!$hotel) {
+                return response("No existe el hotel", 404);
+            } 
+            $hotel->delete(); 
+
+           foreach ($hotel->galeria as $galeria) {        
             $galeria->delete();
             \Storage::delete($galeria->ruta);              
-        }
+            }   
 
-        return response()->json(['message'=>'borrado']);
+            return response("El Hotel ha sido Eliminado", 200);
+        }
+        catch(\Exception $e)
+        {
+            Log::critical("No se puede eliminar el Hotel:{$e->getCode()}, {$e->getLine()}, {$e->getMessage()} ");
+            return response("Alguna cosa esta mal", 500);
+        }
     }
 }
